requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
awscli==1.37.2
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docutils==0.16
eval_type_backport==0.2.2
filelock==3.16.1
Flask==3.1.0
flask-ngrok==0.0.25
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.1
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
propcache==0.2.1
pyarrow==18.1.0
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.7.2
s3transfer==0.11.1
shellingham==1.5.4
six==1.17.0
tabulate==0.9.0
together==1.3.11
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3